Relationship between Product and ProductWebsite and front end improvements
Products now belong to a ProductWebsite, on creation of a Product a
function sets the ProductWebsite's foreign key to the Product. Front-end
improvements meaning that the app now scales to mobile and is more
accesible. Finally started a class that will carry out the regular price
checks on the saved products.