Start to add webdriver feature for scraping information of AJAX sites
When AJAX is used to render things like price, a webdriver is needed
rather than a html parser. Still figuring out whats the best way
to go about this. In the meantime i've added a ajax boolean field to each ProductWebsite to
signify what feature a website needs to extract data.